+++ app/Http/Controllers/CollectionController.php
@@ -27,12 +27,13 @@ class CollectionController extends Controller
     {
         // return Collection::with('Buyer')->paginate(10);
         // dd();
-       return view('pages.Data_Barang.Collection.Master_Collection',
-        [
-            'type_menu'=>'Collection',
-            'Collection'=>Collection::with('Buyer')->filter(request(['search']))->paginate(10)
-        ]
-       );
+        return view(
+            'pages.Data_Barang.Collection.Master_Collection',
+            [
+                'type_menu' => 'Collection',
+                'Collection' => Collection::with('Buyer')->filter(request(['search']))->paginate(10)
+            ]
+        );
     }
 
     /**
@@ -44,7 +45,7 @@ class CollectionController extends Controller
     public function create(RateLimiter $limiter)
     {
         try {
-            if (!in_array(auth()->user()->akses, [1, 2])) {
+            if (!in_array(auth()->user()->akses, [1, 2, 3, 6])) {
                 throw new AuthorizationException();
             }
 
@@ -59,13 +60,17 @@ class CollectionController extends Controller
 
             $limiter->hit($key, $decayMinutes * 60);
 
-            return view('pages.Data_Barang.Collection.Tambah_Collection',
+            // Jika memiliki akses
+
+            return view(
+                'pages.Data_Barang.Collection.Tambah_Collection',
                 [
-                'type_menu'=>'Collection' ,
-                'buyers'=>Buyer::all() 
+                    'type_menu' => 'Collection',
+                    'buyers' => Buyer::all()
                 ]
             );
 
+            // Jika tidak memimiliki akses
         } catch (AuthorizationException $exception) {
             throw new AuthorizationException('Halaman Ini Tidak Boleh Diakses', 403);
         }
@@ -81,16 +86,17 @@ class CollectionController extends Controller
     {
         $validatedData = $request->validate(
             [
-              'id'=>'required',
-              'Nama_Collection' =>'required' ,
-              'Buyer_Id'=>'required'
-            ],[
-                'required'=>'Kolom Tidak Boleh Kosong'
+                'id' => 'required',
+                'Nama_Collection' => 'required',
+                'Buyer_Id' => 'required'
+            ],
+            [
+                'required' => 'Kolom Tidak Boleh Kosong'
             ]
         );
         // return $validatedData;
         Collection::create($validatedData);
-        return redirect('/Collection')->with('success','Data Berhasil Ditambahkan');
+        return redirect('/Collection')->with('success', 'Data Berhasil Ditambahkan');
     }
 
     /**
@@ -101,11 +107,12 @@ class CollectionController extends Controller
      */
     public function show(Collection $Collection)
     {
-        return view('pages.Data_Barang.Collection.Detail_Collection',
+        return view(
+            'pages.Data_Barang.Collection.Detail_Collection',
             [
-                'type_menu'=>'Collection',
+                'type_menu' => 'Collection',
                 'Collection' => $Collection,
-                'items'=> Item::where('Collection_Id',$Collection->id)->get()
+                'items' => Item::where('Collection_Id', $Collection->id)->get()
             ]
         );
     }
@@ -116,14 +123,36 @@ class CollectionController extends Controller
      * @param  \App\Models\Collection  $collection
      * @return \Illuminate\Http\Response
      */
-    public function edit(Collection $Collection)
-    { 
-        
-        return view('pages.Data_Barang.Collection.Edit_Collection',[
-            'type_menu'=>'Collection',
-            'Collection'=>$Collection,
-            'buyers'=>Buyer::all()
-        ]);
+    public function edit(Collection $Collection,RateLimiter $limiter)
+    {
+        try {
+            if (!in_array(auth()->user()->akses, [1, 2, 3, 6])) {
+                throw new AuthorizationException();
+            }
+
+            // Check for brute force attacks
+            $key = 'login.' . request()->ip();
+            $maxAttempts = 5;
+            $decayMinutes = 1;
+
+            if ($limiter->tooManyAttempts($key, $maxAttempts)) {
+                throw new HttpException(Response::HTTP_TOO_MANY_REQUESTS, 'Too many attempts. Please try again later.');
+            }
+
+            $limiter->hit($key, $decayMinutes * 60);
+
+            // Jika memiliki akses
+
+            return view('pages.Data_Barang.Collection.Edit_Collection', [
+                'type_menu' => 'Collection',
+                'Collection' => $Collection,
+                'buyers' => Buyer::all()
+            ]);
+
+            // Jika tidak memimiliki akses
+        } catch (AuthorizationException $exception) {
+            throw new AuthorizationException('Halaman Ini Tidak Boleh Diakses', 403);
+        }
     }
 
     /**
@@ -137,12 +166,13 @@ class CollectionController extends Controller
     {
         $validatedData = $request->validate(
             [
-              'id'=>'required|unique:collections',
-              'Nama_Collection' =>'required' ,
-              'Buyer_Id'=>'required'
-            ],[
-                'required'=>'Kolom Tidak Boleh Kosong',
-                'unique'=>'Kode Telah Digunakan , Silahkan Gunakan Kode Lain'
+                'id' => 'required|unique:collections',
+                'Nama_Collection' => 'required',
+                'Buyer_Id' => 'required'
+            ],
+            [
+                'required' => 'Kolom Tidak Boleh Kosong',
+                'unique' => 'Kode Telah Digunakan , Silahkan Gunakan Kode Lain'
             ]
         );
         // log activity
@@ -156,14 +186,14 @@ class CollectionController extends Controller
             ->withProperties([
                 'old' => $originalData,
                 'new' => $validatedData
-                ])
+            ])
             ->event('Update')
             ->log('This Model has been Update');
 
         //end log activity
         // return $validatedData;
-        Collection::where('id',$Collection->id)->update($validatedData);
-        return redirect('/Collection')->with('success','Data Berhasil Diubah');
+        Collection::where('id', $Collection->id)->update($validatedData);
+        return redirect('/Collection')->with('success', 'Data Berhasil Diubah');
     }
 
     /**
@@ -172,22 +202,45 @@ class CollectionController extends Controller
      * @param  \App\Models\Collection  $collection
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Collection $Collection)
+    public function destroy(Collection $Collection ,RateLimiter $limiter)
     {
-        Collection::destroy($Collection->id);
-        return redirect('/Collection')->with('success','Data Berhasil Dihapus');
+        try {
+            if (!in_array(auth()->user()->akses, [1, 2, 3, 6])) {
+                throw new AuthorizationException();
+            }
+
+            // Check for brute force attacks
+            $key = 'login.' . request()->ip();
+            $maxAttempts = 5;
+            $decayMinutes = 1;
+
+            if ($limiter->tooManyAttempts($key, $maxAttempts)) {
+                throw new HttpException(Response::HTTP_TOO_MANY_REQUESTS, 'Too many attempts. Please try again later.');
+            }
+
+            $limiter->hit($key, $decayMinutes * 60);
+
+            // Jika memiliki akses
+
+            Collection::destroy($Collection->id);
+            return redirect('/Collection')->with('success', 'Data Berhasil Dihapus');
+            
+            // Jika tidak memimiliki akses
+        } catch (AuthorizationException $exception) {
+            throw new AuthorizationException('Halaman Ini Tidak Boleh Diakses', 403);
+        }
     }
 
     public function export()
     {
         // return Collection::with('buyer')->get();
-        return Excel::download(new CollectionExport , 'Collection.xlsx');
+        return Excel::download(new CollectionExport, 'Collection.xlsx');
     }
 
     public function import(Request $request)
     {
-         // Validasi file Excel
-         $request->validate([
+        // Validasi file Excel
+        $request->validate([
             'excel_file' => 'required|mimes:xls,xlsx'
         ]);
